+++ course/views.py
@@ -789,6 +789,12 @@ class HomePage(UserPassesTestMixin, ModelViewSet):
             User.objects.get(username=masquerade) if masquerade else request.user
         )
         courses = self.get_queryset().filter(instructors=user_account)
+        courses = courses.filter(
+            Q(course_term=NEXT_TERM, year=NEXT_YEAR)
+            | Q(course_term=CURRENT_TERM, year=CURRENT_YEAR),
+            course_subject__visible=True,
+            course_schools__visible=True,
+        )
         courses_count = courses.count()
         courses = courses[:15]
         requests = Request.objects.filter(